+++ tests/__init__.py
@@ -1,2 +1,4 @@
 # -*- coding: utf-8 -*-
-from .test_checker import TestChecker
+from .test_amadeus import TestAmadeus, TestAmadeusUtils
+from .test_sirena import TestSirena
+from .test_db import TestDB